refactor Blog and Post into Models and move the document entities into Doctrine/Phpcr
