+++ DependencyInjection/Configuration.php
@@ -51,8 +51,8 @@ class Configuration implements ConfigurationInterface
                     ->end()
                 ->end()
                 ->scalarNode('blog_basepath')
-                    ->isRequired()
                     ->defaultValue('/cms/blog')
+                    ->cannotBeEmpty()
                 ->end()
                 ->scalarNode('routing_post_controller') # unused
                     ->defaultValue('cmf_blog.blog_controller:viewPostAction')
@@ -76,10 +76,10 @@ class Configuration implements ConfigurationInterface
                             ->defaultValue('Symfony\Cmf\Bundle\BlogBundle\Admin\PostAdmin')
                         ->end()
                         ->scalarNode('blog')
-                            ->defaultValue('Symfony\Cmf\Bundle\BlogBundle\Document\Blog')
+                            ->defaultValue('Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Blog')
                         ->end()
                         ->scalarNode('post')
-                            ->defaultValue('Symfony\Cmf\Bundle\BlogBundle\Document\Post')
+                            ->defaultValue('Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Post')
                         ->end()
                     ->end()
                 ->end()
