+++ Resources/config/doctrine-phpcr/Blog.phpcr.xml
@@ -4,18 +4,16 @@
     xsi:schemaLocation="http://doctrine-project.org/schemas/phpcr-odm/phpcr-mapping
         https://github.com/doctrine/phpcr-odm/raw/master/doctrine-phpcr-odm-mapping.xsd">
 
-    <document
-        name="Symfony\Cmf\Bundle\BlogBundle\Document\Blog"
-        referenceable="true"
-        >
+    <document name="Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Blog" referenceable="true">
 
         <id name="id">
-            <generator strategy="PARENT"/>
+            <generator strategy="PARENT" />
         </id>
 
-        <nodename name="name"/>
-        <parent-document name="parent"/>
-        <children name="posts"/>
+        <nodename name="name" />
+        <parent-document name="parent" />
+
+        <children name="posts" />
 
         <referrers name="routes" referring-document="Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr\Route" referenced-by="content">
             <cascade>
@@ -27,4 +25,3 @@
     </document>
 
 </doctrine-mapping>
-
