+++ Resources/config/doctrine-phpcr/Post.phpcr.xml
@@ -5,25 +5,25 @@
         https://github.com/doctrine/phpcr-odm/raw/master/doctrine-phpcr-odm-mapping.xsd">
 
     <document
-        name="Symfony\Cmf\Bundle\BlogBundle\Document\Post"
+        name="Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Post"
         repository-class="Symfony\Cmf\Bundle\BlogBundle\Repository\PostRepository"
         referenceable="true"
         >
 
         <id name="id">
-            <generator strategy="PARENT"/>
+            <generator strategy="PARENT" />
         </id>
 
         <nodename name="name"/>
         <parent-document name="blog"/>
 
-        <field name="slug" type="string"/>
-        <field name="title" type="string"/>
-        <field name="body" type="string"/>
-        <field name="date" type="date"/>
-        <field name="isPublishable" type="boolean"/>
-        <field name="publishStartDate" type="date" nullable="true"/>
-        <field name="publishEndDate" type="date" nullable="true"/>
+        <field name="slug" type="string" />
+        <field name="title" type="string" />
+        <field name="content" type="string" />
+        <field name="publicationDate" type="date" />
+        <field name="isPublishable" type="boolean" />
+        <field name="publishStartDate" type="date" nullable="true" />
+        <field name="publishEndDate" type="date" nullable="true" />
 
         <referrers name="routes" referring-document="Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr\Route" referenced-by="content">
             <cascade>
@@ -35,4 +35,3 @@
     </document>
 
 </doctrine-mapping>
-
