+++ Resources/config/doctrine-phpcr/TagRoute.phpcr.xml
@@ -5,7 +5,7 @@
         https://github.com/doctrine/phpcr-odm/raw/master/doctrine-phpcr-odm-mapping.xsd">
 
       <document
-          name="Symfony\Cmf\Bundle\BlogBundle\Document\TagRoute"
+          name="Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\TagRoute"
           referenceable="true"
           />
