+++ Resources/config/validation.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" ?>
+<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping"
+                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                    xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping
+                                        http://symfony.com/schema/dic/constraint-mapping/constraint-mapping-1.0.xsd">
+
+    <class name="Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Blog">
+        <constraint name="Doctrine\Bundle\PHPCRBundle\Validator\Constraints\ValidPhpcrOdm" />
+
+        <property name="name">
+            <constraint name="NotBlank" />
+        </property>
+
+        <property name="routes">
+            <constraint name="Valid" />
+        </property>
+
+    </class>
+
+    <class name="Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Post">
+        <constraint name="Doctrine\Bundle\PHPCRBundle\Validator\Constraints\ValidPhpcrOdm" />
+
+        <property name="title">
+            <constraint name="NotBlank" />
+        </property>
+
+        <property name="routes">
+            <constraint name="Valid" />
+        </property>
+
+    </class>
+
+</constraint-mapping>
