+++ Resources/doc/building-a-blog.rst
@@ -15,6 +15,6 @@ Register the routes in config.yml
         dynamic:
             ...
             controllers_by_class:
-                Symfony\Cmf\Bundle\BlogBundle\Document\Blog: cmf_blog.blog_controller:listAction
-                Symfony\Cmf\Bundle\BlogBundle\Document\Post: cmf_blog.blog_controller:viewPostAction
-                Symfony\Cmf\Bundle\BlogBundle\Document\Tag: cmf_blog.blog_controller:listAction
+                Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Blog: cmf_blog.blog_controller:listAction
+                Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Post: cmf_blog.blog_controller:viewPostAction
+                Symfony\Cmf\Bundle\BlogBundle\Doctrine\Phpcr\Tag: cmf_blog.blog_controller:listAction
